requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.3
django-admin-interface==0.22.1
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-easy-select2==1.5.8
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-mathfilters==1.0.0
django-modeladmin-reorder==0.3.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
mypy-extensions==0.4.3
numpy==1.24.1
pathspec==0.10.2
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.4
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.12
yarg==0.1.9
django-cors-headers
